requirements.txt
protobuf==3.19.4
grpcio==1.47.0
grpcio-tools==1.44.0
Flask~=2.2.3
gevent~=22.10.2
sentence-transformers~=2.2.2
pandas~=2.0.0
